+++ src/ignitus-Dashboard/ignitus-Dashboard-Header/Containers/dashBoardHeaderContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import dashBoardHeader from '../Components/dashBoardHeader';
-import { logUserOut } from '../../ignitus-Authentication/ignitus-StudentLogin/actions';
+import { logUserOut } from '../../../ignitus-Authentication/ignitus-StudentLogin/actions';
 
 const mapStateToProps = state => ({ state });
 const mapDispatchToProps = dispatch => bindActionCreators({ logUserOut }, dispatch);
