+++ src/ignitus-Dashboard/ignitus-Dashboard-Header/reducers.js
